simplehash api functions id and api key arguments
